2nd version commit

+++ src/main/java/com/alten/testbackend/service/ProductServiceImpl.java
@@ -1,41 +1,68 @@
 package com.alten.testbackend.service;
 
 import com.alten.testbackend.dao.ProductDAO;
+import com.alten.testbackend.dto.ProductDTO;
+import com.alten.testbackend.dto.ProductMapper;
 import com.alten.testbackend.entities.Product;
+import com.alten.testbackend.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class ProductServiceImpl implements ProductService{
 
 
     private ProductDAO productDAO;
+    private ProductMapper mapper;
 
     @Autowired
-    public ProductServiceImpl(ProductDAO productDAO){
+    public ProductServiceImpl(ProductDAO productDAO ,
+                              ProductMapper productMapper){
         this.productDAO = productDAO;
+        this.mapper = productMapper;
     }
     @Override
-    public List<Product> findAll() {
-        return productDAO.findAll();
+    public List<ProductDTO> findAll() {
+
+        return productDAO.findAll()
+                .stream().map(mapper::productToProductDTO)
+                .collect(Collectors.toList());
     }
 
     @Override
-    public Product findById(int id) {
-        return productDAO.findById(id);
+    public ProductDTO findById(int id) {
+
+        Product product = productDAO.findById(id);
+
+        if (product == null) {
+            throw new ResourceNotFoundException ("Product with the given id "+ id +" not found");
+        } else{
+            return mapper.productToProductDTO(productDAO.findById(id));
+        }
+
     }
 
     @Override
     @Transactional
-    public Product save(Product product) {
-        return productDAO.save(product);
+    public ProductDTO save(ProductDTO product) {
+        return mapper.productToProductDTO(productDAO.save(mapper.productDTOToProduct(product)));
     }
 
+
     @Override
+    @Transactional
     public void deleteById(int id) {
-        productDAO.deleteById(id);
+
+        Product productToDelete = productDAO.findById(id);
+
+        if (productToDelete == null) {
+            throw new ResourceNotFoundException ("Product with the given id "+ id +" not found");
+        } else {
+            productDAO.deleteById(id);
+        }
     }
 }
